refactor(mps-sync-plugin-lib): move clases to more fitting packages
