+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/modelix/branch/ModelixBranchListener.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.modelix.mps.sync.modelix
+package org.modelix.mps.sync.modelix.branch
 
 import org.modelix.kotlin.utils.UnstableModelixFeature
 import org.modelix.model.api.IBranch
