+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/modelix/tree/ITreeTraversal.kt
@@ -1,4 +1,4 @@
-package org.modelix.mps.sync.modelix
+package org.modelix.mps.sync.modelix.tree
 
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.coroutineScope
