+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/tasks/SyncQueue.kt
@@ -22,14 +22,14 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import mu.KotlinLogging
 import org.modelix.kotlin.utils.UnstableModelixFeature
-import org.modelix.mps.sync.modelix.BranchRegistry
+import org.modelix.mps.sync.modelix.branch.BranchRegistry
 import org.modelix.mps.sync.mps.notifications.WrappedNotifier
 import org.modelix.mps.sync.mps.util.runReadAction
 import org.modelix.mps.sync.mps.util.toMpsProject
-import org.modelix.mps.sync.transformation.ModelixToMpsSynchronizationException
-import org.modelix.mps.sync.transformation.MpsToModelixSynchronizationException
-import org.modelix.mps.sync.transformation.SynchronizationException
-import org.modelix.mps.sync.transformation.pleaseCheckLogs
+import org.modelix.mps.sync.transformation.exceptions.ModelixToMpsSynchronizationException
+import org.modelix.mps.sync.transformation.exceptions.MpsToModelixSynchronizationException
+import org.modelix.mps.sync.transformation.exceptions.SynchronizationException
+import org.modelix.mps.sync.transformation.exceptions.pleaseCheckLogs
 import org.modelix.mps.sync.util.completeWithDefault
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.ConcurrentHashMap
