+++ mps-sync-plugin-lib/src/main/kotlin/org/modelix/mps/sync/transformation/cache/MpsToModelixMapInitializerVisitor.kt
@@ -10,7 +10,7 @@ import org.modelix.kotlin.utils.UnstableModelixFeature
 import org.modelix.model.api.BuiltinLanguages
 import org.modelix.model.api.IBranch
 import org.modelix.model.api.INode
-import org.modelix.mps.sync.modelix.IBranchVisitor
+import org.modelix.mps.sync.modelix.tree.IBranchVisitor
 import org.modelix.mps.sync.modelix.util.getModel
 import org.modelix.mps.sync.modelix.util.getMpsNodeId
 import org.modelix.mps.sync.modelix.util.nodeIdAsLong
